refactor(index): split keys.rs under the 1000-line cap
keys.rs was 1121 lines. The governance tests — the ones that pin
`symbol_kind_byte`'s u8 ordinals and the length-prefixed key shapes — move
to keys_governance.rs; the encoders stay. Behaviour-preserving: the
ordinals are byte-identical before and after, which matters because
renumbering one would silently miscategorize every persisted entry rather
than fail loudly.
